refactor: Split LayoutService override
